refactor(lns): split lns into core / manipulators / iostream / debug layers (#1334)
Phase 2 continues after cfloat. Same pattern; two differences worth recording.

  header                          lines  headers   I/O
  lns/lns.hpp     (before)        89909      450     5
  lns/core.hpp    (new)           74762      364     0

Criterion 1 met. Criterion 2 not, for the usual reason -- blocktriple.

Preceded by the #1389 sweep over the lns subtree: 30 gaps across 29 headers, fixed
first so the split could be judged on layering rather than transitive breakage.

Out of the core: debugConstexprParameters() and the LnsArithmeticStatistics printer to
a new debug.hpp; operator<< / operator>> to a new iostream.hpp. native/ieee754.hpp
became ieee754_core.hpp -- lns needs only extractFields and ieee754_parameter. The
behavioural switches move into lns_impl.hpp beside the #if blocks that test them.

DIFFERENCE FROM cfloat AND posit: lns defined its stream operators INLINE IN THE CLASS
as non-template friends. An in-class friend definition cannot be moved to another header
as-is, so they are now declared as friend TEMPLATES and defined at namespace scope in
iostream.hpp -- the shape posit's field headers already use.

SPLIT internal/abstract/triple.hpp AS WELL, because it was the binding constraint: after
the lns work its operator<</operator>>/components() still put <iomanip>, <sstream>,
<ostream> and <istream> into the core, leaving 4 I/O-family headers. triple.hpp is
included by SIX impl headers -- lns, takum, dbns, rational, faithful and the
one-parameter skeleton -- so those three text producers were doing that to every one of
their arithmetic cores. The text moves to internal/abstract/triple_io.hpp, the same
split blockbinary, blocksignificand, blockfraction and blocktriple already had (#1394,
#1401, #1403), and triple.hpp drops <iomanip> and switches to ieee754_core.hpp. That is
what takes lns from 4 I/O to 0, and it unblocks takum, which is next in the rollout.

Verified: both gates pass on gcc and clang; lns, takum, dbns and rational suites 67 pass
0 fail; every api.cpp 55 pass 0 fail; clang spot-check 20 pass 0 fail; the hardened
#1389 sweep reports 0 gaps across all 32 lns headers.

Two pre-existing breakages confirmed unrelated and left alone:
skeleton_1param/oneparam_impl.hpp includes a mistyped "universal/native/ieee-754.hpp"
and fails identically at HEAD.
