+++ include/sw/universal/number/lns/math/logarithm.hpp
@@ -5,6 +5,7 @@
 //
 // This file is part of the universal numbers project, which is released under an MIT Open Source license.
 
+#include <cmath>         // std::log
 namespace sw { namespace universal {
 
 // Natural logarithm of x
