+++ include/sw/universal/number/lns/math/minmax.hpp
@@ -6,6 +6,7 @@
 //
 // This file is part of the universal numbers project, which is released under an MIT Open Source license.
 
+#include <algorithm>     // std::min / std::max
 namespace sw { namespace universal {
 
 template<unsigned nbits, unsigned rbits, typename bt, auto... xtra>
